gpio_keys: n1: ignore capacitive button events during screen off
Else home, menu and back button events are reported during
every resume or also when screen is off e.g. in a call

This adds a general config setting for gpio_key
buttons to set this flag so this will not influence
existing gpio_key configs on other devices e.g. find5

Change-Id: I1957ede7a5ef1d0e9d79a24befbaa4d859e0599e